requirements.txt
Pillow==11.0.0
pandas==2.2.3